+++ x/gov/migrations/v4/json.go
@@ -29,6 +29,7 @@ func MigrateJSON(oldState *v1.GenesisState) (*v1.GenesisState, error) {
 		defaultParams.BurnProposalDepositPrevote,
 		defaultParams.BurnVoteQuorum,
 		defaultParams.BurnVoteVeto,
+		defaultParams.MinDepositRatio,
 	)
 
 	return &v1.GenesisState{